+++ packages/client/src/modals/project/index.ts
@@ -3,6 +3,7 @@ export * from './deleteFile';
 export * from './draft';
 export * from './duplicate';
 export * from './loadPrompt';
+export * from './postprocess';
 export * from './rename';
 export * from './tagger';
 export * from './unsaved';