+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- Added a button to easily postprocess (upscale) output images after generation.
 - Added buttons to reset all settings (without resetting the bundle) and reset the bundle (without resetting settings) to "About Metastable".
 - Added keyboard shortcuts (`Up`/`Down`/`Enter`) for selecting models in model browser.
 - Added support for trackpad gestures in image previews.